@@ -1,181 +1,335 @@
 <?php
 
-namespace Jasny;
+declare(strict_types=1);
 
-use Jasny\Auth\User;
-use Jasny\Auth\Middleware;
+namespace Jasny\Auth;
+
+use Jasny\Auth\AuthzInterface as Authz;
+use Jasny\Auth\Confirmation\ConfirmationInterface as Confirmation;
+use Jasny\Auth\Confirmation\NoConfirmation;
+use Jasny\Auth\ContextInterface as Context;
+use Jasny\Auth\Session\PhpSession;
+use Jasny\Auth\Session\SessionInterface as Session;
+use Jasny\Auth\StorageInterface as Storage;
+use Jasny\Auth\UserInterface as User;
+use Jasny\Immutable;
+use Psr\EventDispatcher\EventDispatcherInterface as EventDispatcher;
 
 /**
- * Authentication and access control
- * 
- * <code>
- * class Auth extends Jasny\Auth
- * {
- *     use Jasny\Auth\ByLevel;
- *     use Jasny\Auth\Sessions;
- * 
- *     protected $levels = [
- *       'user' => 1,
- *       'admin' => 10
- *     ];
- * 
- *     public function fetchUserById($id)
- *     {
- *         ...
- *     }
- * 
- *     public function fetchUserByUsername($username)
- *     {
- *         ...
- *     }
- * }
- * </code>
+ * Authentication and authorization.
  */
-abstract class Auth
+class Auth implements Authz
 {
-    /**
-     * Current authenticated user
-     * @var User|false
-     */
-    protected $user;
-    
-    
-    /**
-     * Persist the current user id across requests
-     * 
-     * @return void
-     */
-    abstract protected function persistCurrentUser();
-    
-    /**
-     * Get current authenticated user id
-     * 
-     * @return mixed
-     */
-    abstract protected function getCurrentUserId();
-    
-    /**
-     * Fetch a user by ID
-     * 
-     * @param int|string $id
-     * @return User|null
-     */
-    abstract public function fetchUserById($id);
+    use Immutable\With;
 
     /**
-     * Fetch a user by username
-     * 
-     * @param string $username
-     * @return User|null
+     * Stateful authz service.
+     * A new copy will be set if user is logged in or out, or if context changes.
      */
-    abstract public function fetchUserByUsername($username);
-    
-    
-    /**
-     * Get current authenticated user
-     * 
-     * @return User|null
-     */
-    public function user()
-    {
-        if (!isset($this->user)) {
-            $uid = $this->getCurrentUserId();
-            $this->user = $uid ? ($this->fetchUserById($uid) ?: false) : false;
-        }
-        
-        return $this->user ?: null;
-    }
-    
-    /**
-     * Set the current user
-     * 
-     * @param User $user
-     * @return User|null
-     */
-    public function setUser(User $user)
-    {
-        if ($user->onLogin() === false) {
-            return null;
-        }
-        
-        $this->user = $user;
-        $this->persistCurrentUser();
-        
-        return $this->user;
-    }
-    
-    
-    /**
-     * Hash a password
-     * 
-     * @param string $password
-     * @return string
-     */
-    public function hashPassword($password)
-    {
-        if (!is_string($password) || $password === '') {
-            throw new \InvalidArgumentException("Password should be a (non-empty) string");
-        }
-        
-        return password_hash($password, PASSWORD_BCRYPT);
-    }
-    
-    /**
-     * Fetch user and verify credentials.
-     * 
-     * @param User|null $user
-     * @param string    $password
-     * @return boolean
-     */
-    public function verifyCredentials($user, $password)
-    {
-        return isset($user) && password_verify($password, $user->getHashedPassword());
-    }
-    
-    /**
-     * Login with username and password
-     * 
-     * @param string $username
-     * @param string $password
-     * @return User|null
-     */
-    public function login($username, $password)
-    {
-        $user = $this->fetchUserByUsername($username);
+    protected Authz $authz;
 
-        if (!isset($user) || !$this->verifyCredentials($user, $password)) {
-            return null;
-        }
-        
-        return $this->setUser($user);
-    }
-    
+    protected Session $session;
+    protected Storage $storage;
+    protected Confirmation $confirmation;
+    protected EventDispatcher $dispatcher;
+
+    /** The service can't be used before it's initialized */
+    protected bool $initialized = false;
+
     /**
-     * Logout
+     * Auth constructor.
      */
-    public function logout()
+    public function __construct(Authz $authz, Storage $storage, ?Confirmation $confirmation = null)
     {
-        $user = $this->user();
-        
-        if (!$user) {
+        $this->authz = $authz;
+        $this->storage = $storage;
+        $this->confirmation = $confirmation ?? new NoConfirmation();
+
+        // Set default services
+        $this->session = new PhpSession();
+        $this->dispatcher = self::dummyDispatcher();
+    }
+
+    /**
+     * Get a copy with a different session manager.
+     */
+    public function withSession(Session $session): self
+    {
+        return $this->withProperty('session', $session);
+    }
+
+    /**
+     * Get a copy with an event dispatcher.
+     */
+    public function withEventDispatcher(EventDispatcher $dispatcher): self
+    {
+        return $this->withProperty('dispatcher', $dispatcher);
+    }
+
+
+    /**
+     * Initialize the service using session information.
+     */
+    public function initialize(): void
+    {
+        if ($this->initialized) {
+            throw new \LogicException("Auth service is already initialized");
+        }
+
+        ['uid' => $uid, 'context' => $cid, 'checksum' => $checksum] = $this->session->getInfo();
+
+        $user = $uid !== null ? $this->storage->fetchUserById($uid) : null;
+        $context = $cid !== null ? $this->storage->fetchContext($cid) : null;
+
+        if ($user !== null && $user->getAuthChecksum() !== $checksum) {
+            $user = null;
+            $context = null;
+        }
+
+        $this->authz = $this->authz->forUser($user)->inContextOf($context);
+        $this->initialized = true;
+    }
+
+    /**
+     * Is the service is initialized?
+     */
+    public function isInitialized(): bool
+    {
+        return $this->initialized;
+    }
+
+    /**
+     * Throw an exception if the service hasn't been initialized yet.
+     *
+     * @throws \LogicException
+     */
+    protected function assertInitialized(): void
+    {
+        if (!$this->initialized) {
+            throw new \LogicException("Auth needs to be initialized before use");
+        }
+    }
+
+
+    /**
+     * Get all available authorization roles (for the current context).
+     *
+     * @return string[]
+     */
+    final public function getAvailableRoles(): array
+    {
+        return $this->authz->getAvailableRoles();
+    }
+
+    /**
+     * Check if the current user is logged in and has specified role.
+     *
+     * <code>
+     *   if (!$auth->is('manager')) {
+     *     http_response_code(403); // Forbidden
+     *     echo "You are not allowed to view this page";
+     *     exit();
+     *   }
+     * </code>
+     */
+    final public function is(string $role): bool
+    {
+        $this->assertInitialized();
+
+        return $this->authz->is($role);
+    }
+
+    /**
+     * Get current authenticated user.
+     *
+     * @return User|null
+     */
+    final public function user(): ?User
+    {
+        $this->assertInitialized();
+
+        return $this->authz->user();
+    }
+
+    /**
+     * Get the current context.
+     */
+    final public function context(): ?Context
+    {
+        $this->assertInitialized();
+
+        return $this->authz->context();
+    }
+
+
+    /**
+     * Set the current user.
+     *
+     * @throws LoginException
+     */
+    public function loginAs(User $user): void
+    {
+        $this->assertInitialized();
+
+        if ($this->authz->user() !== null) {
+            throw new \LogicException("Already logged in");
+        }
+
+        $this->loginUser($user);
+    }
+
+    /**
+     * Login with username and password.
+     *
+     * @throws LoginException
+     */
+    public function login(string $username, string $password): void
+    {
+        $this->assertInitialized();
+
+        if ($this->authz->user() !== null) {
+            throw new \LogicException("Already logged in");
+        }
+
+        $user = $this->storage->fetchUserByUsername($username);
+
+        if ($user === null || !$user->verifyPassword($password)) {
+            throw new LoginException('Invalid credentials', LoginException::INVALID_CREDENTIALS);
+        }
+
+        $this->loginUser($user);
+    }
+
+    /**
+     * Set the current user and dispatch login event.
+     *
+     * @throws LoginException
+     */
+    private function loginUser(User $user): void
+    {
+        $event = new Event\Login($this, $user);
+        $this->dispatcher->dispatch($event);
+
+        if ($event->isCancelled()) {
+            throw new LoginException($event->getCancellationReason(), LoginException::CANCELLED);
+        }
+
+        $this->authz = $this->authz->forUser($user);
+
+        $this->updateSession();
+    }
+
+    /**
+     * Logout current user.
+     */
+    public function logout(): void
+    {
+        $this->assertInitialized();
+
+        $user = $this->authz->user();
+
+        if ($user === null) {
+            return; // already logged out
+        }
+
+        $this->authz = $this->authz->forUser(null)->inContextOf(null);
+        $this->updateSession();
+
+        $this->dispatcher->dispatch(new Event\Logout($this, $user));
+    }
+
+    /**
+     * Set the current context.
+     */
+    public function setContext(?Context $context): void
+    {
+        $this->assertInitialized();
+
+        $this->authz = $this->authz->inContextOf($context);
+        $this->updateSession();
+    }
+
+    /**
+     * Recalculate authz roles for current user and context.
+     * Store the current auth information in the session.
+     *
+     * @return $this
+     */
+    public function recalc(): self
+    {
+        $this->authz = $this->authz->recalc();
+        $this->updateSession();
+
+        return $this;
+    }
+
+    /**
+     * Store the current auth information in the session.
+     */
+    protected function updateSession(): void
+    {
+        $user = $this->authz->user();
+        $context = $this->authz->context();
+
+        if ($user === null) {
+            $this->session->clear();
             return;
         }
-        
-        $user->onLogout();
-        
-        $this->user = false;
-        $this->persistCurrentUser();
+
+        $uid = $user->getAuthId();
+        $cid = $context !== null ? $context->getAuthId() : null;
+        $checksum = $user->getAuthChecksum();
+
+        $this->session->persist($uid, $cid, $checksum);
     }
-    
-    
+
+
     /**
-     * Create auth middleware interface for access control.
-     *
-     * @param callable $getRequiredRole
-     * @return Middleware
+     * Return read-only service for authorization of the current user and context.
      */
-    public function asMiddleware($getRequiredRole)
+    public function authz(): Authz
     {
-        return new Middleware($this, $getRequiredRole);
+        return $this->authz;
+    }
+
+    /**
+     * Return read-only service for authorization of the specified user.
+     */
+    public function forUser(?User $user): Authz
+    {
+        return $this->authz->forUser($user);
+    }
+
+    /**
+     * Get an authz service for the given context.
+     */
+    public function inContextOf(?Context $context): Authz
+    {
+        return $this->authz->inContextOf($context);
+    }
+
+
+    /**
+     * Get service to create or validate confirmation token.
+     */
+    public function confirm(string $subject): Confirmation
+    {
+        return $this->confirmation->withStorage($this->storage)->withSubject($subject);
+    }
+
+
+    /**
+     * Create an event dispatcher as null object.
+     * @codeCoverageIgnore
+     */
+    private static function dummyDispatcher(): EventDispatcher
+    {
+        return new class () implements EventDispatcher {
+            /** @inheritDoc */
+            public function dispatch(object $event): object
+            {
+                return $event;
+            }
+        };
     }
 }
